Extracted duplicate code into a base class, as indicated by Sonar.
