+++ src/main/java/com/checkmarx/flow/custom/GitLabSecurityDashboard.java
@@ -27,7 +27,7 @@ import java.util.List;
 @Service("GitLabDashboard")
 @RequiredArgsConstructor
 @Slf4j
-public class GitLabSecurityDashboard implements IssueTracker {
+public class GitLabSecurityDashboard extends ImmutableIssueTracker {
     private static final String ISSUE_FORMAT = "%s @ %s : %d";
 
     private final GitLabProperties properties;
@@ -128,36 +128,6 @@ public class GitLabSecurityDashboard implements IssueTracker {
     @Override
     public Issue createIssue(ScanResults.XIssue issue, ScanRequest request) throws MachinaException {
         return null;
-    }
-
-    @Override
-    public void closeIssue(Issue issue, ScanRequest request) throws MachinaException {
-
-    }
-
-    @Override
-    public Issue updateIssue(Issue issue, ScanResults.XIssue resultIssue, ScanRequest request) throws MachinaException {
-        return null;
-    }
-
-    @Override
-    public String getIssueKey(Issue issue, ScanRequest request) {
-        return issue.getId();
-    }
-
-    @Override
-    public String getXIssueKey(ScanResults.XIssue issue, ScanRequest request) {
-        return issue.getFilename();
-    }
-
-    @Override
-    public boolean isIssueClosed(Issue issue, ScanRequest request) {
-        return false;
-    }
-
-    @Override
-    public boolean isIssueOpened(Issue issue, ScanRequest request) {
-        return false;
     }
 
     @Data
