+++ src/main/java/com/checkmarx/flow/custom/ImmutableIssueTracker.java
@@ -1,0 +1,38 @@
+package com.checkmarx.flow.custom;
+
+import com.checkmarx.flow.dto.Issue;
+import com.checkmarx.flow.dto.ScanRequest;
+import com.checkmarx.flow.exception.MachinaException;
+import com.checkmarx.sdk.dto.ScanResults;
+
+public abstract class ImmutableIssueTracker implements IssueTracker {
+    @Override
+    public void closeIssue(Issue issue, ScanRequest request) throws MachinaException {
+
+    }
+
+    @Override
+    public Issue updateIssue(Issue issue, ScanResults.XIssue resultIssue, ScanRequest request) throws MachinaException {
+        return null;
+    }
+
+    @Override
+    public String getIssueKey(Issue issue, ScanRequest request) {
+        return issue.getId();
+    }
+
+    @Override
+    public String getXIssueKey(ScanResults.XIssue issue, ScanRequest request) {
+        return issue.getFilename();
+    }
+
+    @Override
+    public boolean isIssueClosed(Issue issue, ScanRequest request) {
+        return false;
+    }
+
+    @Override
+    public boolean isIssueOpened(Issue issue, ScanRequest request) {
+        return false;
+    }
+}
